Add command to build clang compilation database
A compilation database is quite handy, since it provides enough
information to reproduce the results of a build independently of the
build system.

This information is typically consumed by applications like clangd and
used to provide source completion, hints and documentation within
popular editors.

On VS Code it is consumed by the C/C++ plugin. On VIM multiple plugins
can be used, being some of the most popular YouCompleteMe or Conquer
of Completion.

As given by the specification of the format:
https://clang.llvm.org/docs/JSONCompilationDatabase.html