Support for static libraries
This is rather stupid, as we don't parallelize the build of multiple
targets yet, but at least we can build multiple libraries link against
them and make sure the build order is correct, as dependencies are
checked and built for every target.

This also means that there is no checks in place for circular
dependencies across targets, which needs to happen in the future.